naive mod polymapper
