+++ pytorch_segmentation_models_trainer/custom_models/mod_polymapper/modpolymapper.py
@@ -1,0 +1,30 @@
+# -*- coding: utf-8 -*-
+"""
+/***************************************************************************
+ pytorch_segmentation_models_trainer
+                              -------------------
+        begin                : 2021-11-16
+        git sha              : $Format:%H$
+        copyright            : (C) 2021 by Philipe Borba - Cartographic Engineer
+                                                            @ Brazilian Army
+        email                : philipeborba at gmail dot com
+ ***************************************************************************/
+/***************************************************************************
+ *                                                                         *
+ *   This program is free software; you can redistribute it and/or modify  *
+ *   it under the terms of the GNU General Public License as published by  *
+ *   the Free Software Foundation; either version 2 of the License, or     *
+ *   (at your option) any later version.                                   *
+ *                                                                         *
+ ****
+"""
+import torch
+from torch import nn
+
+
+class NaiveModPolyMapper(nn.Module):
+    """
+    This class implements a naive mod poly mapper.
+    """
+
+    pass
